requirements.txt
numpy
psycopg2
python-dotenv>=1.0.0